+++ assets/src/common/helpers/index.js
